feat: export loadMiddleware and loadServerMod function from plugin-server
