+++ .changeset/slimy-files-perform.md
@@ -1,0 +1,7 @@
+---
+'@modern-js/plugin-server': patch
+---
+
+feat: export loadMiddleware and loadServerMod function from plugin-server
+
+feat: 导出 loadMiddleware 和 loadServerMod 函数
